Restructure flatten library module, and break intrinsics module out into another file
