+++ Cargo.toml
@@ -7,3 +7,9 @@ authors = [ "James Friend <jamesf@jsdf.co>" ]
 [dependencies]
 itertools = "0.4"
 linenoise-rust = "0.1.14"
+
+[lib]
+name = "lisp"
+
+[[bin]]
+name = "repl"
